Fixed Destruction
The elements are now layered. A ray is sent from each element in different directions. When the beam collides, the distance to the target is checked, and if the distance is not great, it is added to the collection. After enumeration of all elements, all elements added to the collection are destroyed.