Added Constraint
Added limitation of moving elements. Now the element cannot be moved outside the "grid" border. Constraints are determined by the starting position of the grid and the maximum number of cells in the grid.
Fixed mesh normalization. After the swipe, the hanging elements "fall". After the "fall" of each element, a check is made: whether it is possible to destroy the blocks of elements.